Standardized degree value
