+++ encoder.py
@@ -2,21 +2,21 @@ from PIL import Image
 from keygen import *
 from utils import *
 import argparse
+import getpass
 
-
-def encode(image_path,degree,pwd):
+def encode(image_path, pwd):
     extension=image_path.split('.')[-1]
-
     try:
-    	im = Image.open(image_path, "r")
+        im = Image.open(image_path, "r")
     except FileNotFoundError:
-    	print ('Image path is incorrect. Try again.')
-    	exit(0)
+        print ('Image path is incorrect. Try again.')
+        exit(0)
 
     im = Image.open(image_path, "r")
     arr = im.load()  # pixel data stored in this 2D array
     (W, H) = im.size
     print(W, H)
+    degree= int(0.36*W*H)
 
     KEY = generate_tuples(H, W, pwd)
 
@@ -34,14 +34,6 @@ def encode(image_path,degree,pwd):
     return (im,arr,saved_path)
 
 
-def efficiency_calc(image_path,im,arr):
-    im2 = Image.open(image_path, "r")
-    (W,H) = im.size
-    # To calculate efficiency of the algorithm
-    arr2 = im2.load()
-    efficiency(arr2, arr, W, H)
-    im.show()
-
 if __name__=='__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument('--path', help='absolute/relative path to the image file')
@@ -52,7 +44,7 @@ if __name__=='__main__':
     if(image_path is None):
         print ('Please provide the path to image file. Try again.')
         exit(0)
-    degree = int(input("Enter degree: "))
+    # degree = int(input("Enter degree: "))
     pwd = getpass.getpass("Enter password: ")
-    (im,arr,saved_path)=encode(image_path,degree,pwd)
-    efficiency_calc(image_path,im,arr)
+    (im,arr,saved_path)=encode(image_path, pwd)
+    efficiency_calc(image_path,im,arr, saved_path)
